Move annotator.plugin.viewer -> annotator.ui.viewer.standalone
